+++ VERSION_HISTORY.md
@@ -1,3 +1,9 @@
+### v41
+* version updates for `bwa`, `pairix`, `cooler`, `pairsamtools`.
+
+### v40
+* pairsamtools pipeline added
+
 ### v39
 * bwa-mem 0.7.16a-r1185-dirty
 * run-bwa-mem.sh copies input files to output directory, so that unzipping doesn't change input files.